requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.0.0
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.1
oauthlib==3.1.0
Pillow==8.0.1
priority==1.3.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pymongo==3.11.2
pyOpenSSL==20.0.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
zope.interface==5.2.0
